# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /company-settings/{id}/documents/upload-url:
    post:
      operationId: createCompanyBranchDocumentUploadUrl
      tags: [company-branch-documents]
      summary: Create a private upload URL for a company branch document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchDocumentUploadRequest"
      responses:
        "200":
          description: Private upload URL and object path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BranchDocumentUploadUrl"
        "400":
          description: Invalid file metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Active branch not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /company-settings/{id}/documents:
    get:
      operationId: getCompanyBranchDocuments
      tags: [company-branch-documents]
      summary: List documents attached to a company branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Branch documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BranchDocument"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Active branch not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createCompanyBranchDocument
      tags: [company-branch-documents]
      summary: Attach an uploaded document to a company branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchDocumentInput"
      responses:
        "201":
          description: Branch document saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BranchDocument"
        "400":
          description: Invalid document metadata or upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Active branch or uploaded file not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /company-settings/{id}/documents/{documentId}/file:
    get:
      operationId: getCompanyBranchDocumentFile
      tags: [company-branch-documents]
      summary: Download a private company branch document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: The private image or PDF document
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Branch document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
    BranchDocumentUploadRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        contentType:
          type: string
          enum: [application/pdf, image/jpeg, image/png, image/webp]
        size:
          type: integer
          minimum: 1
      required: [name, contentType, size]
    BranchDocumentUploadUrl:
      type: object
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
          pattern: "^/objects/"
      required: [uploadURL, objectPath]
    BranchDocumentInput:
      type: object
      properties:
        document_type:
          type: string
          enum: [tax_number, cr_number, national_address, custom]
        title:
          type: string
          maxLength: 120
        object_path:
          type: string
          pattern: "^/objects/"
        file_name:
          type: string
          maxLength: 255
        content_type:
          type: string
          enum: [application/pdf, image/jpeg, image/png, image/webp]
      required: [document_type, title, object_path, file_name, content_type]
    BranchDocument:
      type: object
      properties:
        id:
          type: integer
        branch_id:
          type: integer
        document_type:
          type: string
          enum: [tax_number, cr_number, national_address, custom]
        title:
          type: string
        file_name:
          type: string
        content_type:
          type: string
          enum: [application/pdf, image/jpeg, image/png, image/webp]
        file_size:
          type: integer
        created_at:
          type: string
      required: [id, branch_id, document_type, title, file_name, content_type, file_size, created_at]
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...